Create a persistent UI window using scriptUI.
Due to Adobe's completely brain dead everything, this requires using Bridge to
send the function to Illustrator as a string so it can be eval'd.

Also change the back end to only convert the currently selected layer to ASS, as
this makes more sense with the new UI. I will probably change it to allow what
is being output to be configured in the near future. And then I will hopefully
never touch this absolutely awful scripting API ever again ever.