Add more collection methods, reformat to match my current standards.
There is now a drop down menu to choose which collection method to use. The
three that are available are: current layer, inner shadow, and all layers. Inner
shadow is a hacky version of current layer that expects the layer to be
formatted in a very specific manner. How to use it will be added to the readme
soon. The other two do exactly what it sounds like they do.

Also, the box that the exported text is written to is now highlighted to make
copying from it easier. How exciting!

A misplaced alert pop up has been squashed.

I will probably end up eating these words but if I ever have to touch this
scripting API again it will be too soon.